+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Requires PHP `8.4`
+- Requires `innmind/server-control:~7.0`
+
 ## 5.0.0 - 2025-05-08
 
 ### Changed